+++ HaruUp/Presentation/History/HistoryCoordinator.swift
@@ -11,14 +11,14 @@ final class HistoryCoordinator: Coordinator {
     let navigationController: UINavigationController
     var childCoordinators: [any Coordinator] = []
     
-    private let historyService: HistoryServiceProtocol = HistoryService()
+    private let missionService: MissionServiceProtocol = MissionService()
     
     init(navigationController: UINavigationController) {
         self.navigationController = navigationController
     }
     
     func start() {
-        let historyVM = HistoryViewModel(historyService: historyService)
+        let historyVM = HistoryViewModel(missionService: missionService)
         let historyVC = HistoryViewController(viewModel: historyVM)
         
         navigationController.setViewControllers([historyVC], animated: false)
